// src/app/shared/interfaces/user.ts
export interface User {
  id: string;
  email: string;
  first_name?: string;
  last_name?: string;
  avatar?: string;
  role?: string;
  date_created?: string;
  date_updated?: string;
  last_access?: string;
}

export interface AuthResponse {
  access_token: string;
  refresh_token: string;
  expires: number;
  user: User;
}

export interface LoginData {
  email: string;
  password: string;
}

export interface RegisterData {
  email: string;
  password: string;
  first_name: string;
  last_name: string;
}